make session module depend on service-rpc instead of jdbc
