+++ jdbc/src/main/java/org/apache/iotdb/jdbc/Utils.java
@@ -27,8 +27,6 @@ import java.util.regex.Pattern;
 import org.apache.iotdb.service.rpc.thrift.TSDataValue;
 import org.apache.iotdb.service.rpc.thrift.TSQueryDataSet;
 import org.apache.iotdb.service.rpc.thrift.TSRowRecord;
-import org.apache.iotdb.service.rpc.thrift.TS_Status;
-import org.apache.iotdb.service.rpc.thrift.TS_StatusCode;
 import org.apache.iotdb.tsfile.exception.write.UnSupportedDataTypeException;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
 import org.apache.iotdb.tsfile.read.common.Field;
@@ -73,17 +71,6 @@ public class Utils {
     }
 
     return params;
-  }
-
-  /**
-   * verify success.
-   *
-   * @param status -status
-   */
-  public static void verifySuccess(TS_Status status) throws IoTDBSQLException {
-    if (status.getStatusCode() != TS_StatusCode.SUCCESS_STATUS) {
-      throw new IoTDBSQLException(status.errorMessage);
-    }
   }
 
   /**
@@ -148,64 +135,4 @@ public class Utils {
     }
   }
 
-
-  public static ByteBuffer getTimeBuffer(RowBatch rowBatch) {
-    ByteBuffer timeBuffer = ByteBuffer.allocate(rowBatch.getTimeBytesSize());
-    for (int i = 0; i < rowBatch.batchSize; i++) {
-      timeBuffer.putLong(rowBatch.timestamps[i]);
-    }
-    timeBuffer.flip();
-    return timeBuffer;
-  }
-
-  public static ByteBuffer getValueBuffer(RowBatch rowBatch) {
-    ByteBuffer valueBuffer = ByteBuffer.allocate(rowBatch.getValueBytesSize());
-    for (int i = 0; i < rowBatch.measurements.size(); i++) {
-      TSDataType dataType = rowBatch.measurements.get(i).getType();
-      switch (dataType) {
-        case INT32:
-          int[] intValues = (int[]) rowBatch.values[i];
-          for (int index = 0; index < rowBatch.batchSize; index++) {
-            valueBuffer.putInt(intValues[index]);
-          }
-          break;
-        case INT64:
-          long[] longValues = (long[]) rowBatch.values[i];
-          for (int index = 0; index < rowBatch.batchSize; index++) {
-            valueBuffer.putLong(longValues[index]);
-          }
-          break;
-        case FLOAT:
-          float[] floatValues = (float[]) rowBatch.values[i];
-          for (int index = 0; index < rowBatch.batchSize; index++) {
-            valueBuffer.putFloat(floatValues[index]);
-          }
-          break;
-        case DOUBLE:
-          double[] doubleValues = (double[]) rowBatch.values[i];
-          for (int index = 0; index < rowBatch.batchSize; index++) {
-            valueBuffer.putDouble(doubleValues[index]);
-          }
-          break;
-        case BOOLEAN:
-          boolean[] boolValues = (boolean[]) rowBatch.values[i];
-          for (int index = 0; index < rowBatch.batchSize; index++) {
-            valueBuffer.put(BytesUtils.boolToByte(boolValues[index]));
-          }
-          break;
-        case TEXT:
-          Binary[] binaryValues = (Binary[]) rowBatch.values[i];
-          for (int index = 0; index < rowBatch.batchSize; index++) {
-            valueBuffer.putInt(binaryValues[index].getLength());
-            valueBuffer.put(binaryValues[index].getValues());
-          }
-          break;
-        default:
-          throw new UnSupportedDataTypeException(
-              String.format("Data type %s is not supported.", dataType));
-      }
-    }
-    valueBuffer.flip();
-    return valueBuffer;
-  }
 }
