+++ jdbc/src/test/java/org/apache/iotdb/jdbc/UtilsTest.java
@@ -26,6 +26,7 @@ import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
+import org.apache.iotdb.rpc.RpcUtils;
 import org.apache.iotdb.service.rpc.thrift.TSDataValue;
 import org.apache.iotdb.service.rpc.thrift.TSQueryDataSet;
 import org.apache.iotdb.service.rpc.thrift.TSRowRecord;
@@ -69,13 +70,13 @@ public class UtilsTest {
   @Test
   public void testVerifySuccess() {
     try {
-      Utils.verifySuccess(new TS_Status(TS_StatusCode.SUCCESS_STATUS));
+      RpcUtils.verifySuccess(new TS_Status(TS_StatusCode.SUCCESS_STATUS));
     } catch (Exception e) {
       fail();
     }
 
     try {
-      Utils.verifySuccess(new TS_Status(TS_StatusCode.ERROR_STATUS));
+      RpcUtils.verifySuccess(new TS_Status(TS_StatusCode.ERROR_STATUS));
     } catch (Exception e) {
       return;
     }
