+++ session/pom.xml
@@ -54,7 +54,12 @@
   <dependencies>
     <dependency>
       <groupId>org.apache.iotdb</groupId>
-      <artifactId>iotdb-jdbc</artifactId>
+      <artifactId>service-rpc</artifactId>
+      <version>0.9.0-SNAPSHOT</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.iotdb</groupId>
+      <artifactId>tsfile</artifactId>
       <version>0.9.0-SNAPSHOT</version>
     </dependency>
   </dependencies>
