+++ session/src/main/java/org/apache/iotdb/session/Config.java
@@ -1,0 +1,9 @@
+package org.apache.iotdb.session;
+
+public class Config {
+
+  public static final String DEFAULT_USER = "root";
+  public static final String DEFALUT_PASSWORD = "root";
+
+
+}
