+++ session/src/main/java/org/apache/iotdb/session/IoTDBSessionException.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.exception;
+package org.apache.iotdb.session;
 
 public class IoTDBSessionException extends RuntimeException {
